--- commerce/interaction-api/src/main/java/ru/yandex/practicum/smarthometech/commerce/api/config/InteractionApiModuleAutoConfiguration.java
@@ -1,9 +1,0 @@
-package ru.yandex.practicum.smarthometech.commerce.api.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-
-@AutoConfiguration
-@ComponentScan("ru.yandex.practicum.smarthometech.commerce.api.mapper")
-public class InteractionApiModuleAutoConfiguration {
-}
