--- commerce/interaction-api/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-ru.yandex.practicum.smarthometech.commerce.api.config.InteractionApiModuleAutoConfiguration
